requirements.txt
flask==3.0.0
flask-cors==4.0.0
websocket-client==1.6.4
gunicorn==21.2.0
